Moving types into separate file to reduce dependency with provisioning_portal.rb.
